tests, checkstyle, rename package
