+++ library/tinkerpop/src/test/java/uk/gov/gchq/gaffer/tinkerpop/process/traversal/strategy/GafferPopGraphStepStrategyTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.gaffer.tinkerpop.process.traversal.stategy;
+package uk.gov.gchq.gaffer.tinkerpop.process.traversal.strategy;
 
 import org.apache.tinkerpop.gremlin.process.traversal.P;
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource;
