+++ library/tinkerpop/src/test/resources/gafferpop-test.properties
@@ -19,3 +19,6 @@ gaffer.storeproperties=src/test/resources/gaffer/store.properties
 gaffer.schemas=src/test/resources/gaffer/schema/
 gaffer.userId=user01
 gaffer.operation.options=key1:value1
+gaffer.elements.getalllimit=2
+gaffer.elements.hasstepfilterstage=POST_AGGREGATION
+
